+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Master (will become release 2.8)
 
+## Deprecations and removals
+
+- Remove `Intersection`'s deprecated enums `dimension` and
+  `codimension`. Instead use grid's dimension and 1.
+
+- Remove deprecated `Grid::getRealImplementation`.
+  Use the `impl()` member function of the facade classes directly instead.
+
+- Remove GeometryGrid's deprecated constructors accepting raw pointers.
+
+- Remove deprecated `AlbertaGrid::getRealIntersection`.
+  Use the `impl()` member function of the intersection class directly instead.
+
+- Remove deprecated `AlbertaGrid::readGridXdr` and `AlbertaGrid::writeGridXdr`.
+  Instead use `AlbertaGrid::readGrid` and `AlbertaGrid::writeGrid`.
+
+- Remove deprecated header `dune/common/universalmapper.hh`.
+
 # Release 2.7
 
 - The `YaspGrid` class has a new constructor that takes a `Coordinates`